Send email after order confirmation and fix multiple image upload

+++ com/main/controller/products/update/UpdateProductController.php
@@ -15,10 +15,12 @@ class UpdateProductController extends Controller {
 
     private $repository;
     private $fileService;
+    private $factory;
     
     public function __construct() {
         $this->repository = new ProductRepository();
         $this->fileService = new FileService();
+        $this->factory = new ProductFactory();
     }
 
     public function canHandle(Request $request) {
@@ -27,12 +29,26 @@ class UpdateProductController extends Controller {
 
     public function handle(Request $request) {
         $id = $request->getPathParameter("product");
+
         $product = $this->retrieveProductById($id);
 
-        $this->fileService->deleteFilesFrom($product->imageNamesAsArray(), Constants::$PRODUCTS_UPLOAD_DIRECTORY);
-        $this->fileService->uploadFilesInto($request->getFiles("productimage"), Constants::$PRODUCTS_UPLOAD_DIRECTORY);
+        $files = $request->getFiles("productimage");
+        if ($files->areUploaded()) {
+            $this->fileService->deleteFilesFrom($product->imageNamesAsArray(), Constants::$PRODUCTS_UPLOAD_DIRECTORY);
+            $this->fileService->uploadFilesInto($files, Constants::$PRODUCTS_UPLOAD_DIRECTORY);
+        }
 
-        $this->repository->update((new ProductFactory())->createProductFromRequest($request));
+        $imageNames = '';
+        if ($files->areUploaded()) {
+            $imageNames = $files->getUniqueNames();
+        } else {
+            $imageNames = $product->imageName();
+        }
+
+        $product = $this->factory->createProductFromRequest($request);
+        $product->setImageName($imageNames);
+
+        $this->repository->update($product);
 
         return (new ResponseBuilder())->withStatusCodeOK()->build();
     }
